Actualziacion uso de sessiones para autenticacion
se agrego el uso de sessiones para poder autorizar el uso de las api

+++ api/routes/admins.js
@@ -44,6 +44,7 @@ router.post('/login', (req, res) => {
       return res.status(400).json({ mensaje: 'No existe el Correo registrado', error: { Error: true } });
     }
 
+    req.session.user = { correo: correo };
     return res.status(201).json({
       mensaje: 'Usuario logeado exitosamente',
       data: { correo, access_control: "si" },
@@ -56,31 +57,35 @@ router.post('/onGuardarBanco', async (req, res) => {
   const { nombre, interes, anios, enganche, correo } = req.body;
   const isAdmin = await esAdmin(correo);
 
-  if (isAdmin) {
-    const query = 'SELECT * FROM Bancos WHERE Nombre = ? AND Interes = ? AND Anios = ? AND Enganche = ?';
-    connection.query(query, [nombre, interes, anios, enganche], (err, results) => {
-      if (err) {
-        return res.status(500).json({ mensaje: 'Error de base de datos', error: { Error: true, error: err } });
-      }
-
-      if (results.length > 0) {
-        return res.status(400).json({ mensaje: 'El banco con las especificaciones ya existe', error: { Error: true } });
-      }
-
-      const query = 'INSERT INTO Bancos (Nombre, Interes, Anios, Enganche) VALUES (?, ?, ?, ?)';
+  if (req.session.user.correo) {
+    if (isAdmin) {
+      const query = 'SELECT * FROM Bancos WHERE Nombre = ? AND Interes = ? AND Anios = ? AND Enganche = ?';
       connection.query(query, [nombre, interes, anios, enganche], (err, results) => {
         if (err) {
-          return res.status(500).json({ mensaje: 'Error al crear el Banco', error: { Error: true, error: err } });
+          return res.status(500).json({ mensaje: 'Error de base de datos', error: { Error: true, error: err } });
         }
 
-        return res.status(201).json({
-          mensaje: 'Banco creado exitosamente',
-          data: { nombre, interes, anios, enganche },
+        if (results.length > 0) {
+          return res.status(400).json({ mensaje: 'El banco con las especificaciones ya existe', error: { Error: true } });
+        }
+
+        const query = 'INSERT INTO Bancos (Nombre, Interes, Anios, Enganche) VALUES (?, ?, ?, ?)';
+        connection.query(query, [nombre, interes, anios, enganche], (err, results) => {
+          if (err) {
+            return res.status(500).json({ mensaje: 'Error al crear el Banco', error: { Error: true, error: err } });
+          }
+
+          return res.status(201).json({
+            mensaje: 'Banco creado exitosamente',
+            data: { nombre, interes, anios, enganche },
+          });
         });
       });
-    });
+    } else {
+      return res.status(403).json({ mensaje: 'No es admin', error: { Error: true } });
+    }
   } else {
-    return res.status(403).json({ mensaje: 'No es admin', error: { Error: true } });
+    return res.status(500).json({ mensaje: 'No session', error: { Error: true } });
   }
 });
 
@@ -89,20 +94,24 @@ router.post('/onEliminarBanco', async (req, res) => {
   const { idBanco, correo } = req.body;
   const isAdmin = await esAdmin(correo);
 
-  if (isAdmin) {
-    const query = 'DELETE FROM Bancos WHERE ID = ?';
-    connection.query(query, [idBanco], (err, results) => {
-      if (err) {
-        return res.status(500).json({ mensaje: 'Error de base de datos', error: { Error: true, error: err } });
-      }
+  if (req.session.user.correo) {
+    if (isAdmin) {
+      const query = 'DELETE FROM Bancos WHERE ID = ?';
+      connection.query(query, [idBanco], (err, results) => {
+        if (err) {
+          return res.status(500).json({ mensaje: 'Error de base de datos', error: { Error: true, error: err } });
+        }
 
-      return res.status(201).json({
-        mensaje: 'Banco eliminado exitosamente',
-        data: { ok: true },
+        return res.status(201).json({
+          mensaje: 'Banco eliminado exitosamente',
+          data: { ok: true },
+        });
       });
-    });
+    } else {
+      return res.status(403).json({ mensaje: 'No es admin', error: { Error: true } });
+    }
   } else {
-    return res.status(403).json({ mensaje: 'No es admin', error: { Error: true } });
+    return res.status(500).json({ mensaje: 'No session', error: { Error: true } });
   }
 });
 
@@ -111,28 +120,32 @@ router.post('/onModificarBancos', async (req, res) => {
   const { bancos, correo } = req.body;
   const isAdmin = await esAdmin(correo);
   let flag = true
-  if (isAdmin) {
-    for (elemento of bancos) {
-      const query = 'UPDATE Bancos SET Nombre = ?, Interes = ?, Anios = ?, Enganche = ? WHERE ID = ?';
-      connection.query(query, [elemento.nombre, elemento.interes, elemento.anios, elemento.enganche, elemento.id], (err, results) => {
-        if (err) {
-          flag = false
-        }
-      });
-
-      if (flag === false) {
-
-        return res.status(500).json({ mensaje: 'Error de base de datos', error: { Error: true } });
-
-      } else {
-        return res.status(201).json({
-          mensaje: 'Bancos actualizados exitosamente',
-          data: { ok: true },
+  if (req.session.user.correo) {
+    if (isAdmin) {
+      for (elemento of bancos) {
+        const query = 'UPDATE Bancos SET Nombre = ?, Interes = ?, Anios = ?, Enganche = ? WHERE ID = ?';
+        connection.query(query, [elemento.nombre, elemento.interes, elemento.anios, elemento.enganche, elemento.id], (err, results) => {
+          if (err) {
+            flag = false
+          }
         });
+
+        if (flag === false) {
+
+          return res.status(500).json({ mensaje: 'Error de base de datos', error: { Error: true } });
+
+        } else {
+          return res.status(201).json({
+            mensaje: 'Bancos actualizados exitosamente',
+            data: { ok: true },
+          });
+        }
       }
+    } else {
+      return res.status(403).json({ mensaje: 'No es admin', error: { Error: true } });
     }
   } else {
-    return res.status(403).json({ mensaje: 'No es admin', error: { Error: true } });
+    return res.status(500).json({ mensaje: 'No session', error: { Error: true } });
   }
 });
 
@@ -141,20 +154,35 @@ router.post('/onEliminarCliente', async (req, res) => {
   const { rfc, correo } = req.body;
   const isAdmin = await esAdmin(correo);
 
-  if (isAdmin) {
-    const query = 'DELETE FROM Clientes WHERE RFC = ?';
-    connection.query(query, [rfc], (err, results) => {
-      if (err) {
-        return res.status(500).json({ mensaje: 'Error de base de datos', error: { Error: true, error: err } });
-      }
+  if (req.session.user.correo) {
+    if (isAdmin) {
+      const query = 'DELETE FROM Clientes WHERE RFC = ?';
+      connection.query(query, [rfc], (err, results) => {
+        if (err) {
+          return res.status(500).json({ mensaje: 'Error de base de datos', error: { Error: true, error: err } });
+        }
 
-      return res.status(201).json({
-        mensaje: 'Cliente eliminado exitosamente',
-        data: { ok: true },
+        return res.status(201).json({
+          mensaje: 'Cliente eliminado exitosamente',
+          data: { ok: true },
+        });
       });
+    } else {
+      return res.status(403).json({ mensaje: 'No es admin', error: { Error: true } });
+    }
+  } else {
+    return res.status(500).json({ mensaje: 'No session', error: { Error: true } });
+  }
+});
+
+router.post('/getSession', async (req, res) => {
+  if (req.session.user !== undefined) {
+    return res.status(201).json({
+      mensaje: 'Session ok',
+      data: { ok: true },
     });
   } else {
-    return res.status(403).json({ mensaje: 'No es admin', error: { Error: true } });
+    return res.status(500).json({ mensaje: 'No session', error: { Error: true } });
   }
 });
